+++ app/src/main/java/camp/nextstep/edu/github/ui/MainActivity.kt
@@ -1,0 +1,34 @@
+package camp.nextstep.edu.github.ui
+
+import android.os.Bundle
+import androidx.activity.viewModels
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.ViewModelProvider
+import camp.nextstep.edu.github.MyApplication
+import camp.nextstep.edu.github.databinding.ActivityMainBinding
+import camp.nextstep.edu.github.di.MainComponent
+import javax.inject.Inject
+
+class MainActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityMainBinding
+    private lateinit var mainComponent: MainComponent
+
+    @Inject
+    lateinit var viewModelFactory: ViewModelProvider.Factory
+    private val viewModel: MainViewModel by viewModels { viewModelFactory }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        mainComponent = (application as MyApplication).appComponent.mainComponent().create()
+        mainComponent.inject(this)
+
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        with(binding) {
+            lifecycleOwner = this@MainActivity
+            vm = viewModel
+            rvSearchResult.adapter = RepositoriesSearchAdapter()
+        }
+    }
+}
